rename websocket server
